Fix SunOS 4.1.4 warnings and add platform documentation
SunOS 4.x compatibility fixes:
- Add extern "C" wrappers for system headers lacking C++ guards
- Add function prototypes for gettimeofday, ioctl, pclose, bcmp
- Add prototypes for socket functions with C linkage
- Fix void* to const char* casts for bcmp calls
- Add explicit casts for double-to-int/float/long conversions
- Undefine conflicting macros between termios.h and sys/ioctl.h
- Add fprintf/strtol/strtod declarations for nxjson.c

CxTime fix:
- Initialize struct tm with memset before use
- Set tm_isdst = -1 to let mktime determine DST automatically

Documentation:
- Add PLATFORM_SUPPORT.md with library/test matrix per platform
- Add "Building on Vintage Unix" section to README

+++ base/double.cpp
@@ -171,19 +171,19 @@ CxDouble::operator double( void )
 
 CxDouble::operator int(void)
 {
-    int iValue = value;
+    int iValue = (int)value;
     return( iValue );
 }
 
 CxDouble::operator float(void)
 {
-    float fValue = value;
+    float fValue = (float)value;
     return( fValue );
 }
 
 CxDouble::operator long(void)
 {
-    long lValue = value;
+    long lValue = (long)value;
     return( lValue );
 }
